Add icons and improve form transitions
This commit introduces custom icons for the `DashboardForm`, `LoginForm`, `NewCustomerForm`, `NewProductForm`, and `NewSupplierForm`. The icons are managed using `ComponentResourceManager` and stored in base64 format within their respective `.resx` files.

Additionally, the `OnFormClosing` method has been implemented in `NewCustomerForm`, `NewProductForm`, and `NewSupplierForm` to ensure that the `DashboardForm` is displayed when these forms are closed, enhancing the user experience and application flow.